Safety guard against git tracking of .gqpatches directory.
git itself MUST NEVER track the .gqpatches directory. If it does, 'git
gq pop' would remove/change the patches directory and destroy all
unapplied patches.
This is more of an assertion against a bug in git-gq, this shouldn't
happen.
If this error happens, the error message shows a simple recipe how to
fix the problem.