In case of a conflict, a file .gqpatches/tmp/DIFF.patch is now created.
This shows the complete patch that failed to apply. This may be useful
when trying to resolve the patch.

When a file to modify was not found, git does not create a reject file
so looking into DIFF.patch is the only way to see want changes failed to
apply.