Backup/Restore of the patch queue is now done with git.
This is a much better way then creating tar files.

Now 'git gq backup' creates a backup of the applied and unapplied
patches.

'git gq restore REV' restores the saved state of the patch queue.

'git gq revert' reverts the git repo of the user to the state when the
backup of the patch queue was made.

'git gq qrepo COMMAND -- OPTIONS' runs git commands on the patch queue
repository.