Use pyproject.toml (PEP 518) to specify and install the project.
This enables the program be be distributed with pypi and installed with
'pip install git-gq'.

In order to be compliant with PEP 518 the script 'git-gq' had to be
renamed to 'git_gq.py'. "git-gq" is then created by 'pip'.

A classical tar.gz for installation with an install script can still be
generated with the script 'administration_tools/mk-tar-dist.sh'.