git gq restore/revert were improved.
This is needed to avoid the 'detached head' state. The reverted branch
can either get the name of the old current branch or a new name by
appending a number to the branch name. Which one is chosen is determined
by option --move-branchname.

Some other improvements were made, it is now ensured that 'git gq
restore' cannot be run when there are uncommitted changes or unknown
files in the patch directory.